The Education of Graphic Designers
"A basic daulity of graphic design becomes apparent when the formation of practitioners is considered: what skills do they need to develop?  Graphic design is both a rational and an artistic activity.  The decision-making process in graphic design alternates between the consideration of objective information and intuitive leaps.  The goal of practitioners should be to base their decisions as much as possible on objective information, but the nature of the field always requires a certain degree of artistic intuition, that is, of decisions made by designers on the basis of experience that is difficult to quantify or explain rationally.  (Graphic design in this case is comparable to marketing or psychoanalysis.  All are activities in which a body of knowledge has to be applied to specific situations that relate to human behavior.)
The balance between artistic and rational elements in the practice of graphic design poses an interesting challenge to design educators, a challenge that calls for the development of visual sophistication and intuitive abilities to express concepts visually, along with a rational capacity for processes of analysis and synthesis.  In addition, graphic designers need skills to listen and interpret the needs and concepts of people in other fields and enough flexibility of mind and visual resources to produce efficient communications.
No school could attempt to deal with all of these requirements in every area of professional practice."
  -Graphic Design: Fine Art or Social Science, Jorge Frascara