Content-Out Layout
"Grids do not exist in a vacuum. They exist in relation to the content. We never start with a grid. We start with an idea which is then translated into a form, a structure."<br>
 - Linda van Deursen<br>
This introduces an interesting article, <a href="http://alistapart.com/article/content-out-layout">Content-Out Layout by Nathan Ford</a>, featured on <a href="http://alistapart.com">A List Apart</a>.